improvements

+++ sam/sam.js
@@ -30,14 +30,26 @@ module.exports.serverDirsList = () =>
     fs.readdirSync(`${process.env.HOME}/.corki/servers`);
 
 
+
+// TODO:
+module.exports.backupServerDirs = () => {
+    // backup
+    require("child_process").execSync(`cp /home/tate/.corki/servers /home/tate/.corki/servers_bkp`);
+};
+
 module.exports.pruneServerDirs = () => {
-    let dirs = module.exports.serverDirsList();
+    // probaly api outage
+    if (Array.from(global.client.guilds).length == 0)
+        return;
+
     const fs = require("fs-extra");
-    dirs.forEach(g => {
+
+    // delete all irrelevant server directories
+    module.exports.serverDirsList().forEach(g => {
         if (!global.client.guilds.get(g)) {
             // edit this to make it move the bad server directory to a backup folder
             fs.removeSync(`${process.env.HOME}/.corki/servers/${g}`);
             console.log("pruned server dir: " + g);
         }
     });
-}
+};
